Rename KeycloakProvider to NukeAuthProvider with enhanced token management
